Use the XDG config directory by default
