+++ go/sendgmail/paths.go
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !xdg
-
 package main
 
 import (
@@ -21,6 +19,8 @@ import (
 	"os"
 	"path/filepath"
 )
+
+var useXDG bool = true
 
 func userHomeDir() string {
 	dir, err := os.UserHomeDir()
@@ -30,10 +30,50 @@ func userHomeDir() string {
 	return dir
 }
 
+func userConfigDir() string {
+	if dir := os.Getenv("XDG_CONFIG_HOME"); dir != "" {
+		return dir
+	}
+	dir, err := os.UserHomeDir()
+	if err != nil {
+		log.Fatalf("Neither $XDG_CONFIG_HOME nor the user home directory is defined: %v.", err)
+	}
+	return filepath.Join(dir, ".config")
+}
+
+func dirExists(path string) bool {
+	st, err := os.Stat(path)
+	return err == nil && st.IsDir()
+}
+
+func fileExists(path string) bool {
+	st, err := os.Stat(path)
+	return err == nil && st.Mode().IsRegular()
+}
+
+func init() {
+	switch {
+	case dirExists(filepath.Join(userConfigDir(), "sendgmail")):
+		useXDG = true
+	case fileExists(filepath.Join(userHomeDir(), ".sendgmail.json")):
+		useXDG = false
+	default:
+		useXDG = true
+	}
+}
+
 func configPath() string {
-	return filepath.Join(userHomeDir(), ".sendgmail.json")
+	if useXDG {
+		return filepath.Join(userConfigDir(), "sendgmail", "config.json")
+	} else {
+		return filepath.Join(userHomeDir(), ".sendgmail.json")
+	}
 }
 
 func tokenPath() string {
-	return filepath.Join(userHomeDir(), ".sendgmail."+sender+".json")
+	if useXDG {
+		return filepath.Join(userConfigDir(), "sendgmail", "token."+sender+".json")
+	} else {
+		return filepath.Join(userHomeDir(), ".sendgmail."+sender+".json")
+	}
 }
