--- go/sendgmail/paths_xdg.go
@@ -1,40 +1,0 @@
-// Copyright 2023 Google Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//go:build xdg
-
-package main
-
-import (
-	"os"
-	"path/filepath"
-)
-
-func userConfigDir() string {
-	if dir := os.Getenv("XDG_CONFIG_HOME"); dir != "" {
-		return dir
-	}
-	if dir := os.Getenv("HOME"); dir != "" {
-		return filepath.Join(dir, ".config")
-	}
-	panic("Neither $XDG_CONFIG_HOME nor $HOME is defined.")
-}
-
-func configPath() string {
-	return filepath.Join(userConfigDir(), "sendgmail", "config.json")
-}
-
-func tokenPath() string {
-	return filepath.Join(userConfigDir(), "sendgmail", "token."+sender+".json")
-}
